requirements.txt
absl-py==1.2.0
asgiref==3.5.2
astunparse==1.6.3
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
colorama==0.4.5
cycler==0.11.0
Django==3.2.16
flatbuffers==22.9.24
fonttools==4.37.4
gast==0.4.0
google-auth==2.12.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.49.1
gunicorn==20.1.0
h5py==3.7.0
idna==3.4
imageio==2.22.1
importlib-metadata==5.0.0
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
networkx==2.6.3
numpy==1.21.6
oauthlib==3.2.1
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
Pillow==9.2.0
protobuf==3.19.6
psutil==5.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.4
PyWavelets==1.3.0
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scipy==1.7.3
seaborn==0.12.0
six==1.16.0
sqlparse==0.4.3
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
tifffile==2021.11.2
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.1
typing-extensions==4.4.0
urllib3==1.26.12
Werkzeug==2.2.2
whitenoise==6.2.0
wrapt==1.14.1
zipp==3.9.0
